+++ src/api/pprint.ts
@@ -1,0 +1,4 @@
+import * as printer from '../printer';
+
+export const prettyPrint = printer.prettyPrint;
+export const prettyPrintingOptions = printer.prettyPrintingOptions;